Add build for themed CSS variables
An initial example of a themed token is component.text. The idea is that
the component.text.themed.light token should appear in the tokens.css,
along with any token that doesn't have "themed" in its path. Then token
values with themed.dark should appear in tokens.dark.css.

By putting "themed" in the component itself means that we centralize
information about a component in one place, regardless of the theme.

The name/rsp/themedKebab name transform removes the themed.xyz component
from the naame so that both the "light" and "dark" themed tokens have
the same name. This allows the dark theme to override variables from the
default/light theme.