Crop imagotypes corresponding to "N" clearance
The original imagotypes had fairly large artboards that make it hard to
use these assets out-of-the-bock. In Illustrator and Photoshop I've
re-cropped the imagotypes so that the clearance around the imagotype
corresponds to the height of the Rubin "N", as specific in the Visual
Identity Manual. Note that this clearance is only for when the
imagotype is used alone; in other contexts more space is needed.