Add intermediate-scale imagotype assets
Use these for imagotype heights between 50 px and 109 px.

These files have artboards that are scaled so there is the height "N"
space around the artwork, providing the minimum clearlance necessary for
this imagotype.

The PNG files have heights of 220px, which should scale well in all
situations where these imagotypes are used.