Add cropped versions of rubin imagotypes
The original imagotypes came from the Communications/EPO team with
built-in margin that assured the minimal margin usage guidelines. In
many scenarios though we don't want this built-in margin so that the
logo can align with other content; like in a div column.

These versions are built by cropping the artboard from the SVG original
using Illustrator's bound-to-artwork functionality and then exporting
new SVG and PNG8 renders.